# optimize conversions to native-python uint32 etc. primitive integer types