running the coin selection test from `chia-blockchain`.

with this patch:
```
246.12s call     tests/wallet/test_wallet_trade_store.py::test_large_trade_record_query
190.06s call     tests/wallet/test_transaction_store.py::test_large_tx_record_query
47.61s call     tests/wallet/test_coin_selection.py::TestCoinSelection::test_knapsack_coin_selection
```

with chia_rs 0.4.0:

```
241.49s call     tests/wallet/test_wallet_trade_store.py::test_large_trade_record_query
185.95s call     tests/wallet/test_transaction_store.py::test_large_tx_record_query
14.47s call     tests/wallet/test_coin_selection.py::TestCoinSelection::test_knapsack_coin_selection
```